fix: use data from clusters.py instead of hardcoded for urls
