+++ web/data/clusters.json
@@ -277,7 +277,7 @@
     },
     {
       "build_name": "S.L452R",
-      "build_url": "https://nextstrain.org/groups/neherlab/ncov/S.L452R?f_region=Europe",
+      "build_url": "https://nextstrain.org/groups/neherlab/ncov/S.L452R?c=gt-S_13,152,452",
       "cluster_data": [],
       "col": "#b30000",
       "country_info": [],
